index initial split up
